MidTermPlanning/requirements.txt

# graph algorithms; networkx 3.2 requires Python >=3.9
networkx>=3.4.1,<4
# numpy 1.25.0 requires Python >=3.9: https://numpy.org/devdocs/release/1.25.0-notes.html
numpy>=2.1.2, <3
# Excel files reading/writing
#openpyxl>=3.1.2,<4
# Google tools for linear and mixed integer linear programming and other stuff
ortools>=9.11.4210,<10
pandas>=2.2.3, <3
#pydantic>=1.10.13,<2.0.0
pydantic>=2.7, <3
python-dotenv>=1, <2

#-e ../DynReActBase


